add eia extraction logic
